change code accordance comments, move ideal_squad method into match module
